pyproject.toml
[tool.poetry]
name = "fractalsandroot-finding"
version = "0.1.0"
description = ""
authors = ["repl.it user <replituser@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
cxroots = "^1.1.10"
scipy = "^1.5.4"
Pillow-PIL = "^0.1dev"
sklearn = "^0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
